+++ src/index.ts
@@ -1,3 +1,4 @@
+#!/usr/bin/env node
 import * as crypto from 'crypto';
 import * as fs from 'fs/promises';
 import * as os from 'os';